approval_agent/requirements.txt
python-dotenv
pika
Flask
